+++ client/src/pages/reports/Show.js
@@ -12,6 +12,7 @@ import ConfirmDelete from "components/ConfirmDelete"
 import { ReadonlyCustomFields } from "components/CustomFields"
 import * as FieldHelper from "components/FieldHelper"
 import Fieldset from "components/Fieldset"
+import { parseHtmlWithLinkTo } from "components/editor/LinkAnet"
 import LinkTo from "components/LinkTo"
 import Messages from "components/Messages"
 import { DEFAULT_CUSTOM_FIELDS_PARENT } from "components/Model"
@@ -46,7 +47,6 @@ import { toast } from "react-toastify"
 import { deserializeQueryParams } from "searchUtils"
 import Settings from "settings"
 import utils from "utils"
-import { parseHtmlWithLinkTo } from "utils_links"
 import AuthorizationGroupTable from "./AuthorizationGroupTable"
 import ReportPeople from "./ReportPeople"
